MAY 10, 2017
Into the motherland
The German army march
In the Soviet Union summer 1943
Tanks line up in thousands as far the eye can see
Ready for the onslaught
Ready for the fight
Waiting for the axis to march into a trap
Mines are placed in darkness
In the cover of the night
Waiting to be triggered
When the time is right
Imminent invasion, imminent attack
Once the battle started
There's no turning back
The end of the third Reich draws near
It's time has come to an end
The end of an era is here
It's time to attack!
Into the motherland the German army march
Comrades stand side by side to stop the Nazi charge
Panzers on Russian soil a thunder in the east
One million men at war
Soviet wrath unleashed!
Fields of Prokhorovka
Where the heat of battle burned
Suffered heavy losses
And…